Stop any attack animation buffer when sliding
There are some problems that need to be fixed when sliding. In PlayerMovement, we want the input events "if (Input.GetKeyUp)" to happen in the "update" method instead of inside the "fixed update". Then fixed update needs to have only "RB2D.velocity" movements but are only activated when the input events are true.