Refactor player and add mid air attack system.
Refactor PlayerController, PlayerAttack, and PlayerMovement to ensure that the controller is the only class with logic i.e., using the other classes. Also add player mid air attack system, but needs to be hitting and damaging the enemy too. The player also has multiple animation controllers to have better and cleaner to look at animation transitions.