Add dynamic projectile on hit effect, and refactor code.
Within the CharacterJump animator controller, you can see all the states in the animator tab. With this in mind, some of the bow attack states are set as "attack ender" which means that particular bow attack animation produces a projectile that can pierce through enemies or GameObjects. If it is not an "attack ender" it cannot pierce through enemies. This results in big damage for pierce and small damage for non-pierce projectile.

Refactor some of the currently used code.