Major fixes and updates to combat
NEW ADDITIONS

Add skill system (code wise, animation wise, and how it interacts with the enemy), but it is incomplete as it is not fully integrated to the game i.e., player cannot obtain this naturally at runtime and must be given to the player in the editor itself.

Add 1 new skill for the sword weapon - Skill slash. It slashes vertically 2 times, then a horizontal slash that also moves the player forward in the facing direction.

Add jump boost effects for all skills (bow, sword) when their skills are activated.

Add a base enemy that has many prefab variants i.e., prefab variants are Skeleton, Gobli, etc. which may have their own unique components.

Add damage number system that shows the damage dealt by the player to the enemy.

Add an animator controller component (class) that changes the animator controller of an animator. This is used by both the player and enemy. It is flawed, code wise, in that the enum values are the same for both the player and enemy, meaning the enemy could get and change into an animator controller that is not meant for the enemy.

Add abstract movement class to have efficient code structure and code reusability. Incomplete as the enemy movement class is empty but surely uses the parent's method to move - the class should have its own methods too.

Add targeting system component (class) that, when attached to a gameobject of either a player or enemy, registers a gameobject as their target(s). This class replaces all enemies' SearchTarget class that has similar responsibilities.

Add better UI design i.e., health bar, energy bar, damage numbers, texts, skill icons, etc. from Itch.io.

Add 1 sound from Freesound.org that is used when the player uses Player Air Sword Attack 3 Landed.

REFACTOR

The air combo attacks are split from 1 animator controller to 3 animator controller with the first one having the original air combo attacks, the second and third animator controllers having more combo attacks allowed.

Each skill now have their own class and is a child of the Skill class (abstract) where they must inherit the Effect method. This method is then called by the player controller. This allows for abstraction and for easy to add on to.

The PlayerAttack class' attack methods (bow and sword) now has their coroutine methods removed and replaced to just a normal method as coroutine methods are no longer needed. This was initially used for adding attack rates for the player, however, the attack rate will now be applied to the attack animation speed instead.

Since new UI design is added, the text class is changed from TextMeshPro to just the normal Text class. This is applied to all classes that previously used TextMeshPro for displaying a text.

Encapsulated most classes in namespaces. These namespaces are named and structured the same way as the structure of the Scripts folder.

Rename some of the classes' variables to the proper C# naming conventions i.e., private variables having "_" at the start of the name.

FIXES

Fix player from being stuck on the platform edges when running and colliding with them.

Fix player not being able to collide with the items.

Fix player being able to get the item effect twice when it only should be 1.

Fix player being able to infinitely attack mid air when its head it touching the ceiling.

Fix player being able to input attacks while in the final sword air attack. This causes the player to stay mid air forever should they keep pressing and if they have a large energy capacity.

Fix bow air attacks (pierce) not having any pushback effect that it should have. This is fixed by changing the transition to the exit state to have an exit time of 1 second and transition duration of 0.

Fix grounded attacks not applying the proper pushback effect it should have, it used to only push the enemy vertically, now it pushes them vertically and horizontally if the pushback effect has horizontal pushback.

Fix grounded attacks having lingering pushback effects when its pushback effect should be back to 0 after its animation state ends.