Add equipment system (incomplete) part 2
Can now create a random equipment with proper probability distribution instead of the simple if and else if statement for creating a chance. It is still incomplete as it does not have the Legendary special stat and epic special stat, it also does not hook it up to the game yet.