Add equipment system (incomplete) part 4
Refactored the stats to have 2 kinds of stats, flat or percent. When we get percent we also increase the bonus which is calculated with the rating, making the equipment have a higher rating than an equipment with pure flat bonuses. It is still incomplete as we need to have logic for legendary special effect, and it is not hooked up properly (i.e., how does this interact with the game) in the game.