Add equipment system (incomplete) part 11
Reworked damage numbers to display just above the health bar and removed any physics to it.

Added reward system when the level clear condition is met.

Added epic stats effect to be the same stat as the first stat but is percentage based instead.

Added UI for Equipment system to allow viewing and removing equipment.

Added cap to Critical Chance, Damage Reduction and Evasion Rate stats.

Now need to rework playerpickup class to add to the MyStat instead. Also need to hook up the atk spd to the aerial attacks. And lastly, need to add this equipment system for levels 1 and 3.